+++ .run/ShopBackendApiAppApplication.run.xml
@@ -1,0 +1,11 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="ShopBackendApiAppApplication" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot">
+    <module name="shop-backend-api-app" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="com.petrov.shopbackendapiapp.ShopBackendApiAppApplication" />
+    <option name="ALTERNATIVE_JRE_PATH" />
+    <option name="SHORTEN_COMMAND_LINE" value="NONE" />
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
+  </configuration>
+</component>
