+++ shop-admin-app/src/main/java/com/petrov/controller/LoginController.java
@@ -1,13 +1,7 @@
 package com.petrov.controller;
 
-import com.petrov.persist.RoleRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-
-import java.util.stream.Collectors;
 
 @Controller
 public class LoginController {
