+++ shop-admin-app/src/main/java/com/petrov/controller/dto/RoleDto.java
@@ -1,4 +1,4 @@
-package com.petrov.controller;
+package com.petrov.controller.dto;
 
 import java.util.Objects;
