--- shop-admin-app/src/main/resources/templates/fragments/paging_users.html
@@ -1,14 +1,0 @@
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-
-<a class="page-link" th:fragment="paging (button, page)"
-   th:href="@{/users(
-                        page=${page},
-                        size=${users.size},
-                        sort=${param.sort},
-                        direction=${param.direction},
-                        titleFilter=${param.titleFilter},
-                        minAgeFilter=${param.minAgeFilter},
-                        maxAgeFilter=${param.maxAgeFilter})}"
-   tabindex="-1">[[${button}]]</a>
-
-</html>
