+++ shop-admin-app/src/main/resources/templates/users.html
@@ -41,13 +41,13 @@
             <table class="table table-bordered my-2">
                 <thead>
                 <tr>
-                    <th scope="col" th:replace="fragments/sort_header_users :: sort_header ('Id', 'id')">
+                    <th scope="col" th:replace="fragments/sort_header :: sort_header_user ('Id', 'id')">
                     </th>
 
-                    <th scope="col" th:replace="fragments/sort_header_users :: sort_header ('User name', 'username')">
+                    <th scope="col" th:replace="fragments/sort_header :: sort_header_user ('User name', 'username')">
                     </th>
 
-                    <th scope="col" th:replace="fragments/sort_header_users :: sort_header ('Age', 'age')">
+                    <th scope="col" th:replace="fragments/sort_header :: sort_header_user ('Age', 'age')">
                     </th>
                     <th scope="col">Actions</th>
                 </tr>
@@ -79,7 +79,7 @@
 
                     <li th:if="${users.hasPrevious()}" class="page-item">
                         <a class="page-link"
-                           th:replace="fragments/paging_users :: paging ('Previos', ${users.getNumber()})"></a>
+                           th:replace="fragments/paging :: paging_user ('Previos', ${users.getNumber()})"></a>
                     </li>
                     <li th:unless="${users.hasPrevious()}" class="page-item disabled">
                         <a class="page-link">Previous</a>
@@ -90,12 +90,12 @@
                         class="page-item">
                         <a class="page-link"
                            th:text="${num}"
-                           th:replace="fragments/paging_users :: paging (${num}, ${num})"></a>
+                           th:replace="fragments/paging :: paging_user (${num}, ${num})"></a>
                     </li>
 
                     <li th:if="${users.hasNext()}" class="page-item">
                         <a class="page-link"
-                           th:replace="fragments/paging_users :: paging ('Next', ${users.getNumber()+2})"></a>
+                           th:replace="fragments/paging :: paging_user ('Next', ${users.getNumber()+2})"></a>
                     </li>
                     <li th:unless="${users.hasNext()}" class="page-item disabled">
                         <a class="page-link">Next</a>
