+++ shop-database/src/main/java/com/petrov/persist/RoleRepository.java
@@ -1,5 +1,6 @@
 package com.petrov.persist;
 
+import com.petrov.persist.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
