+++ shop-database/src/main/java/com/petrov/persist/UserSpecifications.java
@@ -1,5 +1,6 @@
 package com.petrov.persist;
 
+import com.petrov.persist.model.User;
 import org.springframework.data.jpa.domain.Specification;
